pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["chrisemke <chrisemke@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "=3.12.2"
Django = "3.1.5"
psycopg2-binary = "^2.9.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
